rename data_availability to blob_consistency
