+++ aggregator/src/blob_consistency/avail.rs
